Fix truncated backprop through time when set on LightningModule and not Trainer (#8804)
* Fix truncated backprop through time set on LightningModule and not Trainer

(cherry picked from commit c4a1c8ba20975ecf5d3ffb76b8cbc107ab761356)
